# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API PDV",
    "description": "Api de controle de ponto de venda",
    "version": "1.5.4"
  },
  "servers": [
    {
      "url": "https://pdv-backend-psi.vercel.app/",
      "description": "URL de produção"
    }
  ], 
  "tags": [
    {"name": "user"},
    {"name": "product"},
    {"name": "category"},
    {"name": "sales"} 
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Exibir todos os users",
        "tags": ["user"]
      }
    },
    "/user/:id": {
      "get": {
        "tags": ["user"],
        "summary": "Procurar o user pela id",
        "parameters": [
          {
            "name": "Bearer Token",
            "in": "Authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/register" : {
      "post": {
        "summary": "Criar usuários",
        "tags": ["user"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string min 8 length"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "logar usuário",
        "description": "Vai verificar os dados e se for encontrado vai retorna que está autorizado",
        "tags": ["user"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string min 8 length"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/edit/:id": {
      "put": {
        "summary": "editar usuário",
        "description": "Rota para editar dados do usuário",
        "tags": ["user"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string min 8 length"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/delete/:id": {
      "delete": {
        "summary": "Deletar um usuário",
        "tags": ["user"]
      }
    },
    "/products": {
      "get": {
        "summary": "Recebe uma Array de todos os produtos",
        "tags": ["product"]
      }
    },
    "/products/:code": {
      "get": {
        "summary": "Recebe o produto pelo código",
        "tags": ["product"]
      }
    },
    "/products/register" : {
      "post": {
        "summary": "Criar produto",
        "tags": ["product"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "code": {
                    "type": "number min length 12"
                  },
                  "name": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/edit/:code" : {
      "put": {
        "summary": "Editar produto",
        "tags": ["product"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "category": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/delete/:code": {
      "delete": {
        "summary": "Deletar um produto",
        "tags": ["product"]
      }
    },
    "/categories": {
      "get": {
        "summary": "Recebe uma Array de todas as categorias",
        "tags": ["category"]
      }
    },
    "/categories/register" : {
      "post": {
        "summary": "Criar categoria",
        "tags": ["category"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/edit/:id" : {
      "put": {
        "summary": "Editar categoria",
        "tags": ["category"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/delete/:id": {
      "delete": {
        "summary": "Deletar uma categoria",
        "tags": ["category"]
      }
    },
    "/sales/today": {
      "get": {
        "summary": "Mostra todas as vendas do dia",
        "tags": ["sales"]
      }
    },
    "/sales/create": {
      "post": {
        "summary": "Registrar uma venda",
        "tags": ["sales"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "charge": {
                    "type": "number"
                  },
                  "paid": {
                    "type": "number"
                  },
                  "paymentMethod": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sales/delete/id": {
      "delete": {
        "summary": "Deletar uma venda por id",
        "tags": ["sales"]
      }
    }
  }
}